+++ bittytax/conv/parsers/__init__.py
@@ -25,6 +25,7 @@ from . import exodus
 from . import gatehub
 from . import gravity
 from . import handcash
+from . import helium
 from . import hitbtc
 from . import hotbit
 from . import ii